+++ azure-pipelines.yml
@@ -16,9 +16,9 @@ steps:
     version: 6.0.x
 
 - task: NodeTool@0
-  displayName: 'Use Node 18.16.x'
+  displayName: 'Use Node 18.16.0'
   inputs:
-    versionSpec: 18.16.x
+    versionSpec: 18.16.0
 
 - task: DotNetCoreCLI@2
   displayName: Publish
