--- src/CaesHelp/.prettierrc
@@ -1,16 +1,0 @@
-﻿{
-"arrowParens": "avoid",
-"bracketSpacing": true,
-"endOfLine": "lf",
-"htmlWhitespaceSensitivity": "css",
-"jsxBracketSameLine": false,
-"jsxSingleQuote": true,
-"printWidth": 80,
-"proseWrap": "preserve",
-"quoteProps": "as-needed",
-"semi": true,
-"singleQuote": true,
-"tabWidth": 2,
-"trailingComma": "none",
-"useTabs": false
-}
