+++ src/CaesHelp/ClientApp/package.json
@@ -1,0 +1,58 @@
+{
+  "name": "caeshelp",
+  "version": "2.0.0",
+  "description": "CAES Help Requests",
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "debug": "ASPNETCORE_ENVIRONMENT=Development dotnet watch run",
+    "devpack": "set NODE_OPTIONS=--openssl-legacy-provider && webpack-dev-server --config webpack.config.js --hot",
+    "webpack": "node node_modules/webpack/bin/webpack.js --env.dev --watch",
+    "lint": "eslint ClientApp --ext .tsx,.ts,.js,.jsx"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/ucdavis/CaesHelp.git"
+  },
+  "author": "Jason Sylvestre",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/ucdavis/CaesHelp/issues"
+  },
+  "dependencies": {
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-dropzone": "^11.0.1",
+    "react-scripts": "5.0.1"
+  },
+  "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^27.5.2",
+    "@types/jquery": "^3.3.36",
+    "@types/node": "^16.18.31",
+    "@types/react": "^18.2.6",
+    "@types/react-dom": "^18.2.4",
+    "@types/ws": "^8.5.4",
+    "eslint-config-react-app": "^7.0.1",
+    "file-loader": "^6.0.0",
+    "isomorphic-fetch": "2.2.1",
+    "sass": "^1.62.1"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ],
+    "rules": {
+      "no-restricted-globals": "warn"
+    }
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not op_mini all"
+  ]
+}
