+++ src/CaesHelp/ClientApp/src/components/InputArray.tsx
@@ -1,0 +1,54 @@
+﻿import * as React from 'react';
+
+export interface IInputArrayProps {
+  handleAddInput: () => void;
+  handleRemoveInput: (idx: any) => void;
+  handleChange: (idx: any, evt: any) => void;
+  name: string;
+  placeholder: string;
+  addButtonName: string;
+  inputs: [{ value: string; isValid: boolean }];
+}
+
+export default class InputArray extends React.Component<IInputArrayProps, any> {
+  public render() {
+    return (
+      <div>
+        <div className="form-group">
+          {this.props.inputs.map((input, idx) => (
+            <div className="input-group" key={idx}>
+              <input
+                className={`form-control ${input.isValid === false &&
+                  'is-invalid'}`}
+                type="text"
+                placeholder={this.props.placeholder}
+                value={input.value}
+                onChange={evt => this.props.handleChange(idx, evt)}
+                name={`${this.props.name}[${idx}]`}
+              />
+              <div className="input-group-append">
+
+                <a
+                  className="btn btn-danger"
+                  type="button"
+                  onClick={() => this.props.handleRemoveInput(idx)}
+                >
+                  <i className="fa fa-times" />
+                </a>
+              </div>
+              <div className="invalid-feedback">This value is invalid</div>
+            </div>
+          ))}
+
+          <button
+            type="button"
+            onClick={() => this.props.handleAddInput()}
+            className="btn btn-primary"
+          >
+            <i className="fa fa-plus" /> {this.props.addButtonName}
+          </button>
+        </div>
+      </div>
+    );
+  }
+}
