+++ src/CaesHelp/Views/Shared/Components/DynamicScripts/Default.cshtml
@@ -1,0 +1,6 @@
+@model CaesHelp.Views.Shared.Components.DynamicScripts.DynamicScriptModel
+
+@foreach (var script in Model.Scripts)
+{
+    @Html.Raw(script)
+}
