+++ src/CaesHelp/Views/Shared/Components/DynamicStyles/Default.cshtml
@@ -1,0 +1,6 @@
+@model string[]
+
+@foreach (var style in Model)
+{
+    @Html.Raw(style)
+}
