+++ src/CaesHelp/Views/Shared/_Layout.cshtml
@@ -20,10 +20,13 @@
         <!-- <link rel="stylesheet" href="https://ucdcdn.azureedge.net/public/ucdavis.1.0.7.css" /> -->
     </environment>
 
-    <!-- <environment names="Staging,Production">
-        <link rel="stylesheet" href="~/dist/vendor.min.css" asp-append-version="true" />
-        <link rel="stylesheet" href="~/dist/app.min.css" asp-append-version="true" />
-    </environment> -->
+    <environment names="Staging,Production">
+        @* In a production deployment, we will use a dynamic view component to load the generated styles *@
+        <cache expires-after="@TimeSpan.FromMinutes(30)">
+            @await Component.InvokeAsync("DynamicStyles")
+        </cache>
+    </environment>
+
     @RenderSection("Styles", required: false)
 </head>
 <body>
@@ -85,6 +88,19 @@
 
     </footer>
 
+    <environment include="Development">
+        @* NOTE: These are auto generated by CRA/ClientApp and will load in our SASS/CSS plus other libraries and will
+            fast reload changes.
+        *@
+        <script src="/static/js/bundle.js"></script>
+    </environment>
+    <environment names="Staging,Production">
+        @* In a production deployment, we will use a dynamic view component to load the generated scripts *@
+        <cache expires-after="@TimeSpan.FromMinutes(30)">
+            @await Component.InvokeAsync("DynamicScripts")
+        </cache>
+    </environment>
+
     <environment names="Development">
         <script type="text/javascript" src="https://code.jquery.com/jquery-3.3.1.js"></script>
         <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.12.9/umd/popper.js"></script>
@@ -95,9 +111,6 @@
         <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.3/umd/popper.min.js" integrity="sha384-ZMP7rVo3mIykV+2+9J3UJ46jBk0WLaUAdn689aCwoqbBJiSnjAK/l8WvCWPIPm49" crossorigin="anonymous"></script>
         <script type="text/javascript" src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
     </environment>
-
-    <script src="~/dist/vendor.js" asp-append-version="true"></script>
-    <script src="~/dist/main.js" asp-append-version="true"></script>
 
     @RenderSection("Scripts", required: false)
 </body>
