--- src/CaesHelp/package.json
@@ -1,76 +1,0 @@
-{
-  "name": "caeshelp",
-  "version": "2.0.0",
-  "description": "CAES Help Requests",
-  "scripts": {
-    "start": "ASPNETCORE_ENVIRONMENT=Development dotnet watch run",
-    "debug": "ASPNETCORE_ENVIRONMENT=Development dotnet watch run",
-    "devpack": "set NODE_OPTIONS=--openssl-legacy-provider && webpack-dev-server --config webpack.config.js --hot",
-    "webpack": "node node_modules/webpack/bin/webpack.js --env.dev --watch",
-    "test": "karma start",
-    "lint": "eslint ClientApp --ext .tsx,.ts,.js,.jsx"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ucdavis/CaesHelp.git"
-  },
-  "author": "Jason Sylvestre",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/ucdavis/CaesHelp/issues"
-  },
-  "homepage": "https://github.com/ucdavis/CaesHelp#readme",
-  "dependencies": {
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-dropzone": "^11.0.1"
-  },
-  "devDependencies": {
-    "@hot-loader/react-dom": "^16.8.6",
-    "@types/jquery": "^3.3.36",
-    "@types/react": "^16.9.34",
-    "@types/react-dom": "^16.9.7",
-    "@types/webpack-env": "^1.15.2",
-    "@typescript-eslint/eslint-plugin": "^2.31.0",
-    "@typescript-eslint/parser": "^2.31.0",
-    "awesome-typescript-loader": "^5.2.1",
-    "babel-eslint": "^10.1.0",
-    "css-loader": "^3.5.3",
-    "eslint": "^6.8.0",
-    "eslint-config-react-app": "^5.2.0",
-    "eslint-plugin-flowtype": "^4.6.0",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.18.3",
-    "eslint-plugin-react-hooks": "^4.0.0",
-    "file-loader": "^6.0.0",
-    "isomorphic-fetch": "2.2.1",
-    "json-loader": "^0.5.7",
-    "mini-css-extract-plugin": "^0.9.0",
-    "sass": "^1.62.1",
-    "optimize-css-assets-webpack-plugin": "^5.0.3",
-    "react-hot-loader": "^4.12.21",
-    "sass-loader": "^8.0.2",
-    "style-loader": "^1.2.1",
-    "typescript": "^3.8.3",
-    "uglifyjs-webpack-plugin": "^2.2.0",
-    "url-loader": "^4.1.0",
-    "webpack": "^4.43.0",
-    "webpack-bundle-analyzer": "^3.7.0",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-middleware": "^3.7.2",
-    "webpack-dev-server": "^3.11.2",
-    "webpack-hot-middleware": "^2.24.3"
-  },
-  "eslintConfig": {
-    "extends": [
-      "eslint:recommended",
-      "plugin:react/recommended",
-      "plugin:@typescript-eslint/recommended",
-      "react-app"
-    ],
-    "rules": {
-      "@typescript-eslint/interface-name-prefix": 0
-    }
-  }
-}
