--- src/CaesHelp/tsconfig.json
@@ -1,22 +1,0 @@
-{
-  "compilerOptions": {
-    "allowJs": true,
-    "baseUrl": ".",
-    "lib": [
-      "dom",
-      "es2015"
-    ],
-    "module": "es2015",
-    "moduleResolution": "node",
-    "target": "es5",
-    "jsx": "react",
-    "sourceMap": true,
-    "skipDefaultLibCheck": true,
-    "noEmit": true
-  },
-  "exclude": [
-    "bin",
-    "node_modules",
-    "wwwroot"
-  ]
-}
